continue commiting historical changes: fix acceleration in char movements, e.g. the char being tossed up vertically. Thing is that those accel movements are approximately simulated, by changing the speed every frame in a linear manner, AND add a simple (speed * frame step) to position. So position is the area sum of a staircase shaped curve instead of real linear. when frame step lowered to 0.5, the area actually is smaller/bigger since the speed gets changed after a 0.5-step frame and in the next half frame the changed speed makes the sum of 2 half frame smaller or bigger than the 1 frame movement.
The current fix is to mark the half frames as odd and even ones---starting when
a move begin, and use the 3c offset dword of the char obj as the flag (it is some padding data).
Then in the odd half we don't apply accelerating factor to the speed,
but do the full applying on the even half. That way the exact old
behavior is achieved.

The current fix may not be the best. For example we can actually do this
when the speed/accel factors are setup, and apply a shift there, so the
result being the same. but later...